move course planning page to `pages` dir
No functionality change.
