+++ resources/js/pages/CoursePlanningPage/components/CourseTable/CourseTable.vue
