+++ resources/js/pages/CoursePlanningPage/components/CourseTable/CourseTableCell.vue
