+++ resources/js/pages/CoursePlanningPage/components/CourseTable/CourseTableCourseRow.vue
