+++ resources/js/pages/CoursePlanningPage/components/CourseTable/CourseTableLeavesRow.vue
