+++ resources/js/pages/CoursePlanningPage/components/CourseTable/EnrollmentInPublishedSection.vue
