+++ resources/js/pages/CoursePlanningPage/components/CourseTable/EnrollmentInUnpublishedSection.vue
