+++ resources/js/pages/CoursePlanningPage/components/CourseTable/index.ts
