+++ resources/js/pages/CoursePlanningPage/components/EditDraftSectionModal.vue
