+++ resources/js/pages/CoursePlanningPage/components/PersonTable/PersonTable.vue
