+++ resources/js/pages/CoursePlanningPage/components/PersonTable/PublishedSectionDetails.vue
