+++ resources/js/pages/CoursePlanningPage/components/PersonTable/UnpublishedSectionDetails.vue
