+++ resources/js/pages/CoursePlanningPage/helpers/getCourseTableRows.ts
