+++ resources/js/pages/CoursePlanningPage/helpers/getEnrollmentsForCourseInTerm.ts
