+++ resources/js/pages/CoursePlanningPage/helpers/getEnrollmentsForPersonInTerm.ts
