+++ resources/js/pages/CoursePlanningPage/helpers/getJoinedEnrollmentRecord.ts
