+++ resources/js/pages/CoursePlanningPage/helpers/getLeavesInTerm.ts
