+++ resources/js/pages/CoursePlanningPage/helpers/getListOfTermLeaves.ts
