+++ resources/js/pages/CoursePlanningPage/helpers/getPersonTableRows.ts
