+++ resources/js/pages/CoursePlanningPage/helpers/serializedCoursePlanningFilters.ts
