+++ resources/js/pages/CoursePlanningPage/helpers/toCourseSpreadsheetRowRecord.ts
