+++ resources/js/pages/CoursePlanningPage/stores/useCoursePlanningStore.ts
