+++ resources/js/pages/CoursePlanningPage/stores/useCourseStore.ts
