+++ resources/js/pages/CoursePlanningPage/stores/useEnrollmentStore.ts
