+++ resources/js/pages/CoursePlanningPage/workers/coursePlanningWorker.ts
