+++ resources/js/pages/CoursePlanningPage/workers/messageTypes.ts
